# speed up `predict_model()`