Rewrites dplyr in `predict_model()` to use vctrs and tibble. :) With `main` dev tune:

``` r
library(tidymodels)

bench::mark(
  fit =  fit_resamples(linear_reg(), mpg ~ ., bootstraps(mtcars)),
  tune = tune_grid(boost_tree(mode = "regression", trees = tune()), mpg ~ ., bootstraps(mtcars)),
  check = FALSE
)
#> # A tibble: 2 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 fit           1.11s    1.11s     0.903      27MB     8.13
#> 2 tune          5.28s    5.28s     0.189     394MB     6.25
```

With this PR:

``` r
#> # A tibble: 2 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 fit           1.05s    1.05s     0.948    26.9MB     8.53
#> 2 tune          5.17s    5.17s     0.193     403MB     5.61
```

<sup>Created on 2023-03-12 with [reprex v2.0.2](https://reprex.tidyverse.org)</sup>